[Improved] Simplify the logic for adding extra pages to the timeline
Allows a developer to explicitly specify the max number of extra pages that a user can have on the main activity.

Please keep in mind the drawbacks to this as well.

Currently there is no foreground changes to accomodate adding more than two pages. This will come in a future commit.