[New] Select any combination of main fragments for the main activity.
Allows you to specify up to 6 swipable pages on the main activity. They are completely custom, so if you do not want a home timeline page or a direct message page, you do not have to use them.

This isn't completely done yet. What isn't working:
- The main drawer array adapter and highlighting the correct current page
- it doesn't take you to the correct mentions page from a notification or the toast
- Wont work well with Blur yet
- Probably some other things I need to find

Be weary if you merge this.