[Fix] Be able to retweet a protected account's retweet.
Case: If your friend has a protected account and he retweets something you thought was funny, before, if you would have tried to retweet it, it would have failed because he (or she) had a protected account.

It will now grab and retweet the original tweet instead of the retweeted one.